refactor(checker): extract direct global type helper
